Separate view from results processing
